Revert "SE-2553 (#16)"
This reverts commit df1a4598bc6dffa3e140264d6af9aaba06e29761.
